pyproject.toml
[tool.poetry]
name = "truthbrush"
version = "0.2.5"
description = "API client for Truth Social"
authors = ["R. Miles McCain <github@sendmiles.email>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.scripts]
truthbrush = "truthbrush.cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.0"
loguru = "^0.7.0"
python-dotenv = "^1.0.1"

python-dateutil = "^2.9.0"
curl_cffi = "^0.14.0"

[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
pytest = "^9.0.1"
pylint = "^4.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
